Smoother transition between different camera usage.  Smoother transition between algorithms.  Also, framecount=0 need not be in the stopCamera

+++ OpenCVB/CameraOakD.vb
@@ -111,8 +111,7 @@ Public Class CameraOakD
         MyBase.GetNextFrameCounts(IMU_FrameTime)
     End Sub
     Public Sub stopCamera()
-        OakProcess.Kill()
+        If OakProcess IsNot Nothing Then OakProcess.Kill()
         pipelineClosed = True
-        frameCount = 0
     End Sub
 End Class
